requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pymysql>=1.1.0
python-dotenv>=1.0.0
pydantic>=2.5.0
hypothesis>=6.92.0
pytest>=7.4.0
httpx>=0.25.0
